Added sponsorController.js file
